requirements.txt
flask
torchvision
torch
numpy
Pillow
tqdm
matplotlib
scikit-learn
